glusterfs: Use a configmap for the heketi topology
The previous implementation was using "oc exec" for writing
the topology file to the deployment pod. I've experienced runs of
openshift-ansible that were stuck because the "cat" process in the
deployment pod was blocked on stdin.

Instead of using "oc exec", mount the topology file to the deployment pod
using a configmap.

Signed-off-by: gbenhaim <galbh2@gmail.com>